requirements.txt
requests==2.32.3
pyyaml==6.0.2
click==8.1.7
tqdm==4.66.6
python-dateutil==2.9.0
pandas==2.2.3
selenium==4.26.1
webdriver-manager==4.0.2
schedule==1.2.2
beautifulsoup4==4.12.3